+++ data/src/main/java/com/moneymong/moneymong/data/datasource/member/MemberRemoteDataSourceImpl.kt
@@ -1,6 +1,7 @@
 package com.moneymong.moneymong.data.datasource.member
 
-import com.moneymong.moneymong.network.api.MoneyMongApi
+import com.moneymong.moneymong.network.api.AgencyApi
+import com.moneymong.moneymong.network.api.MemberApi
 import com.moneymong.moneymong.network.request.member.MemberBlockRequest
 import com.moneymong.moneymong.network.request.member.UpdateAuthorRequest
 import com.moneymong.moneymong.network.response.member.InvitationCodeResponse
@@ -8,25 +9,26 @@ import com.moneymong.moneymong.network.response.member.MemberListResponse
 import javax.inject.Inject
 
 class MemberRemoteDataSourceImpl @Inject constructor(
-    private val moneyMongApi: MoneyMongApi
+    private val memberApi: MemberApi,
+    private val agencyApi: AgencyApi
 ) : MemberRemoteDataSource {
     override suspend fun getInvitationCode(agencyId: Long): Result<InvitationCodeResponse> {
-        return moneyMongApi.getInvitationCode(agencyId)
+        return agencyApi.getInvitationCode(agencyId)
     }
 
     override suspend fun reInvitationCode(agencyId: Long): Result<InvitationCodeResponse> {
-        return moneyMongApi.reInvitationCode(agencyId)
+        return agencyApi.reInvitationCode(agencyId)
     }
 
     override suspend fun getMemberLists(agencyId: Long): Result<MemberListResponse> {
-        return moneyMongApi.getMemberLists(agencyId)
+        return memberApi.getMemberLists(agencyId)
     }
 
     override suspend fun updateMemberAuthor(agencyId : Long, data: UpdateAuthorRequest): Result<Unit> {
-        return moneyMongApi.updateMemberAuthor(agencyId,data)
+        return memberApi.updateMemberAuthor(agencyId,data)
     }
 
     override suspend fun blockMemberAuthor(agencyId: Long, data: MemberBlockRequest): Result<Unit> {
-        return moneyMongApi.blockMemberAuthor(agencyId, data)
+        return memberApi.blockMemberAuthor(agencyId, data)
     }
 }
